Fix some problems writing ZSTD fseq files for layouts with well over 2M channels with complex effects that don't compress well.  Update to newer ZSTD API's on Linux/OSX.

+++ xLights/FSEQFile.h
@@ -84,9 +84,9 @@ public:
     virtual void dumpInfo(bool indent = false);
 
 
-    uint32_t      getNumFrames() const { return m_seqNumFrames; }
+    uint64_t      getNumFrames() const { return m_seqNumFrames; }
     int           getStepTime() const { return m_seqStepTime; }
-    uint32_t      getChannelCount() const  { return m_seqChannelCount; }
+    uint64_t      getChannelCount() const  { return m_seqChannelCount; }
     int           getVersionMajor() const { return m_seqVersionMajor; }
     int           getVersionMinor() const { return m_seqVersionMinor; }
     uint64_t      getUniqueId() const { return m_uniqueId; }
